chore(infra): remove rollup.d.ts
