+++ packages/rolldown/src/plugin/index.ts
@@ -16,7 +16,7 @@ import type { PluginContext } from './plugin-context'
 import type { TransformPluginContext } from './transform-plugin-context'
 import type { NormalizedOutputOptions } from '../options/normalized-output-options'
 import type { LogLevel } from '../log/logging'
-import type { RollupLog } from '../rollup'
+import type { RollupLog } from '../types/misc'
 import type { MinimalPluginContext } from './minimal-plugin-context'
 import type { InputOptions, OutputOptions } from '..'
 import type { BuiltinPlugin } from '../builtin-plugin/constructors'
