+++ packages/rolldown/src/plugin/plugin-driver.ts
@@ -2,7 +2,7 @@ import { getLogHandler, normalizeLog } from '../log/logHandler'
 import { LOG_LEVEL_DEBUG, LOG_LEVEL_INFO, LOG_LEVEL_WARN } from '../log/logging'
 import { Plugin } from './'
 import { error, logPluginError } from '../log/logs'
-import { RollupError } from '../rollup'
+import { RollupError } from '../types/misc'
 import { normalizeHook } from '../utils/normalize-hook'
 import { InputOptions, OutputOptions, RolldownPlugin, VERSION } from '..'
 import { getLogger, getOnLog } from '../log/logger'
