--- packages/rolldown/src/rollup-types.ts
@@ -1,12 +1,0 @@
-export type {
-  OutputOptions,
-  InputOptions,
-  Plugin,
-  OutputPlugin,
-  OutputChunk,
-  NormalizedInputOptions,
-  OutputAsset,
-  OutputBundle,
-  SourceMapInput,
-  RenderedChunk,
-} from './rollup'
